+++ streampipes-sinks-databases-jvm/src/main/java/org/apache/streampipes/sinks/databases/jvm/mysql/Column.java
@@ -57,7 +57,7 @@ class Column {
             case "time":
             case "timestamp":
             case "year":
-                this.type = Datatypes.Long;
+                this.type = Datatypes.Float;
                 def = System.currentTimeMillis();
                 break;
             default:
